Use boxproc abstraction. (#1484)
* In FIR, we want to wrap function pointers in a special box known as a
boxproc value. Fortran has a limited form of dynamic scoping
[https://tinyurl.com/2p8v2hw7] between "host procedures" and "internal
procedures". There are a number of implementations possible.

Boxproc typed values abstract away the implementation details of when a
function pointer can be passed directly (as a raw address) and when a
function pointer has to account for the presence of a dynamic scope.
When lowering Fortran syntax to FIR, all function pointers are emboxed
as boxproc values.

When creating LLVM IR, we must strip away the abstraction and produce
low-level LLVM "assembly" code. This patch implements that
transformation as converting the boxproc values to either raw function
pointers or executable trampolines on the stack as needed. The
trampoline then captures the dynamic scope context within an executable
thunk that can be passed instead of the function's raw address.

Some extra handling is required for Fortran functions that return a
character value to deal with LEN values here.

Add tests for both Burnside and Tilikum bridges.

* Add mlir tests.

* Fix clang-format issue.

* Remove the use of pointer (meant in a generic sense) as it is
overloaded with the Fortran POINTER attribute and may be unintentionally
confusing.

* Handle function signatures with boxproc args.

* Removes todo test.

* Add tests as XFAIL and to be enabled later.