[flang] Validate DIM argument for reduction intrinsics that return a scalar
Reduction intrinsics that have an ARRAY argument of rank 1 can only accept a DIM argument whose value is 1.  We were not detecting cases where DIM was not equal to 1.  Here's a small program that illustrates the problem:
  program minval_test
    integer :: array(3) = [1,2,3]
    integer :: dim
    dim = 0
    print *, minval(array, dim)
  end program minval_test

Our implementation of reduction intrinsics has two sets of interfaces -- one set for intrinsics that return an array and another set for intrinsics that return a scalar.  Some of the intrinsics that return a scalar stem from calls without the DIM argument.  Others stem from calls with the DIM argument where the ARRAY argument has rank 1.  But for scalar-returning calls, we were not passing along the information of whether the call contained a DIM argument.

I fixed this by adding a boolean argument to the interface to the runtime that indicates whether the original source contained a DIM argument.  This small change propatated through a lot of the runtime resulting in several files changing.

This is an initial implementation that only implements this change for the MINVAL intrinsic.  If I receive initial comments validating this approach, I'll expand it to the other reduction intrinsics.