[lowering] lower runtime type info as linkonce objects
A number of issues were caused by derived type descriptors not being
generated as expected, and the issues being only caught at compile time.

This changes ensures lowering will generate runtime type info for all
the derived types that have been seen (converted) in lowering, instead
of relying on the runtime type info to be generated.
The runtime type info will also always be fully defined, even if the
type is not owned by the current compilation unit (e.g. defined in an
external module). To allow this to work without any conflict, the
`linkonce` LLVM linkage will be used. This allows same runtime type info
from different compilation units to be merged at linktime. The property
that a derived type descriptor address uniquely identifies a type will
thus still hold true.

The main motivation of this change is to allow KIND parameterized
derived type to work properly. The derived type descriptor of KIND
parametrized derived type can only be built during the type instantiation.
If the previous approach was kept, this could lead to "diamond" conflicts
if the same kind instantiation of a kind parametrized type defined in a
module A were done in modules B and C, and that a program P used both B and C.
The same runtime type info would be defined in B an C as external, and
caused link time errors when linking P. Using linkonce solves this
issue.

There are two other benefits to this change:
- It is no longer necessary to link against a module that defines
derived types only since the related type descriptors will be generated
by all compilation units that may use it.
- LLVM can safely discard unused runtime time info in the current
compilation unit since it is no longer necessary to retain it in case
another compilation unit might need it. This means that if no
compilation units turns out to need some module runtime type info, it
will just be thrown away by LLVM.